Kernel K2: move update, i18n engine and charts into kernel/
Second slice of the kernel extraction. Unlike K1 this one carries real
logic, so it is gated on behaviour rather than byte-identity.

Seams (deliberately per-module — only two modules need config, so a
kernel-wide init would create false coupling and import-order traps):

- kernel/src/app.ts: configureApp({appId, appName, manifestUrl}), called
  once at the top of main.ts. Consumed by update (manifest URL + the
  manifest 'app' check) and save (window-title suffix, save-picker
  label — branding literals that would otherwise have been silently
  frozen as 'Bento Slides' for every future app).
- kernel/src/i18n.ts: the ENGINE moves; catalogs stay in slides/src/i18n/*
  and register via registerI18n(). slides/src/i18n.ts becomes a facade
  that registers then re-exports, so all 14 consumers are untouched and
  ES module evaluation order guarantees registration precedes the first
  t().
- FIXES a latent bug in the process: locale resolution ran at module
  scope ('let current = resolve()'), and resolve() consults the catalog
  registry. Post-move that would have run against an empty registry and
  silently fallen every non-English viewer back to English. Resolution is
  now lazy and the memo is cleared on registration.
- kernel/src/charts.ts: retyped from ChartElement onto a structural
  ChartLike {w, h, option} — the only fields the engine reads. Slides'
  ChartElement satisfies it, so call sites are unchanged.

PUBLIC_KEY_JWK stays in the kernel as the platform release key; the
manifest 'app' field is what prevents cross-app substitution.

Verified: kernel + slides typecheck clean; conformance gate passes;
SEEDS=100 rig ALL PASS with sync untouched (0 files changed); the built
shell still contains every frozen string verbatim (manifest app id + URL,
pubkey, ' — Bento Slides' suffix, bento/enc, bento-doc) and its
#bento-doc block is byte-identical to shipped v1.0.8. Browser: boots,
window.bento surface unchanged, charts render, updates report 1.0.8, ja
catalog + x-pseudo work, and a zh-TW boot (an alias existing only in the
registered map) renders Traditional Chinese — proving catalogs register
before first use. Zero console errors.
