+++ slides/src/main.ts
@@ -5,6 +5,7 @@
 
 import './styles.css'
 import { anim } from './anim'
+import { configureApp } from '../../kernel/src/app.ts'
 import {
   capturePristine, readEmbeddedDoc, serializeFile, serializeAuto, downloadFile,
   suggestedFileName, parseEnvelope, decryptEnvelope, setEncryptionPassword,
@@ -19,6 +20,14 @@ import { Editor } from './editor/editor'
 import { startPresentation } from './present'
 import { SyncSession } from './sync/session'
 import { onlineTransport, startSharing, stopSharing } from './sync/online'
+
+// Tell the kernel who this app is — must precede any kernel module use
+// (window title suffix, save-picker label, update manifest + its `app` check).
+configureApp({
+  appId: 'bento-slides',
+  appName: 'Bento Slides',
+  manifestUrl: 'https://bento.page/releases/slides/manifest.json',
+})
 
 capturePristine()
